Add `DefaultFriction` and `DefaultRestitution`, rework things
